requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.2.0
alembic==1.14.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2
appnope==0.1.2
APScheduler==3.11.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
anthropic==0.49.0
asttokens==2.0.5
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
Babel==2.11.0
beautifulsoup4==4.12.3
bleach==4.1.0
blinker==1.6.2
Bottleneck==1.4.2
Brotli==1.0.9
CacheControl==0.14.2
cachetools==5.3.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
comm==0.2.1
contourpy==1.2.0
cryptography==41.0.3
cycler==0.11.0
datasets==3.2.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
et_xmlfile==2.0.0
evaluate==0.4.3
exceptiongroup==1.2.0
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.16.1
firebase-admin==6.6.0
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.4.0
git-filter-repo==2.47.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.5.2
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h5py==3.12.1
html5lib==1.1
httpcore==1.0.2
httplib2==0.22.0
httpx==0.27.0
huggingface_hub==0.28.0
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.0
jupyter_server==2.14.1
jupyter_server_terminals==0.4.4
jupyterlab==4.2.5
jupyterlab-pygments==0.1.2
jupyterlab_server==2.27.3
jupyterlab-widgets==3.0.10
keras==3.8.0
Keras-Preprocessing==1.1.2
keras-tuner==1.4.7
kiwisolver==1.4.4
kt-legacy==1.0.5
libclang==18.1.1
lxml==5.3.0
Mako==1.3.8
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
narwhals==1.14.3
nbclient==0.8.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.3
numexpr==2.10.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.61.1
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.13.0
optuna==4.1.0
overrides==7.4.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pillow==10.4.0
pip==24.2
platformdirs==3.10.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.43
proto-plus==1.26.0
protobuf>=5.26.1,<6.0.0
psutil==5.9.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==18.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.9.0
pyOpenSSL==23.2.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-telegram-bot==21.10
pytz==2024.1
PyYAML==6.0.2
pyzmq==25.1.2
referencing==0.30.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.10.6
rsa==4.7.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.2
setuptools==75.1.0
six==1.16.0
streamlit==1.40.2
tensorflow>=2.18.0,<2.19
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem>=0.34.0,<0.37.0
termcolor==2.1.0
tqdm==4.67.0
traitlets==5.14.3
transformers==4.47.1
urllib3==2.2.3
yarl==1.11.0
google-cloud-texttospeech==2.25.1
pydub==0.25.1
python-dotenv==1.1.0
livekit-agents>=1.1.1
livekit-plugins-openai
livekit-plugins-silero
python-dotenv
livekit-api
flask[async]
flask-cors
uvicorn

